contains_regex: move check_contains_regex to new module
The function is re-exported under the old name, so this is transparent
to users of the crate.
